Remove bloated ptrace flags
I have had a think, and these flags are pointless. Since this is a
microkernel, there should always be exactly one way to do something, and
it should be the most flexible and simplest to implement. And you can
already just set breakpoints on the address provided by the signal
handler, and we'll give the entrypoint as an address to the fexec call.